Build Travis for every branch but deploy only on master

+++ deploy/deploy.sh
@@ -1,18 +1,27 @@
-rm updatesite/* -rf
-mkdir updatesite
+#!/bin/bash
 
-cp -r eSourceTrail.site/target/repository/* updatesite/
-cp deploy/index.html updatesite/
+if [[ $TRAVIS_PULL_REQUEST == "false" ]] && [[ $TRAVIS_BRANCH == "master" ]]
+then
+	echo "Deploying now."
+	
+	rm updatesite/* -rf
+	mkdir updatesite
 
-# config
-git config --global user.email "noreply@coati.io"
-git config --global user.name "Travis CI"
+	cp -r eSourceTrail.site/target/repository/* updatesite/
+	cp deploy/index.html updatesite/
 
-git remote add upstream "https://$GH_TOKEN@github.com/$GH_REPO.git"
-git fetch upstream
-git reset upstream/gh-pages
-git add -A updatesite
-git commit -m "Deploy p2 Updatesite"
-git push -q upstream HEAD:gh-pages
+	# config
+	git config --global user.email "noreply@coati.io"
+	git config --global user.name "Travis CI"
+
+	git remote add upstream "https://$GH_TOKEN@github.com/$GH_REPO.git"
+	git fetch upstream
+	git reset upstream/gh-pages
+	git add -A updatesite
+	git commit -m "Deploy p2 Updatesite"
+	git push -q upstream HEAD:gh-pages
+else
+	echo "Not deploying this branch."
+fi
